// opencdx-dashboard-new/src/api/communications/model/email-template.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Timestamp } from './timestamp';

/**
 * 
 * @export
 * @interface EmailTemplate
 */
export interface EmailTemplate {
    /**
     * 
     * @type {string}
     * @memberof EmailTemplate
     */
    'templateId'?: string;
    /**
     * 
     * @type {string}
     * @memberof EmailTemplate
     */
    'subject'?: string;
    /**
     * 
     * @type {string}
     * @memberof EmailTemplate
     */
    'content'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof EmailTemplate
     */
    'variables'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof EmailTemplate
     */
    'templateType'?: TemplateType;
    /**
     * 
     * @type {Timestamp}
     * @memberof EmailTemplate
     */
    'created'?: Timestamp;
    /**
     * 
     * @type {Timestamp}
     * @memberof EmailTemplate
     */
    'modified'?: Timestamp;
    /**
     * 
     * @type {string}
     * @memberof EmailTemplate
     */
    'creator'?: string;
    /**
     * 
     * @type {string}
     * @memberof EmailTemplate
     */
    'modifier'?: string;
}

export const TemplateType = {
    TemplateTypeUnspecified: 'TEMPLATE_TYPE_UNSPECIFIED',
    TemplateTypeNotification: 'TEMPLATE_TYPE_NOTIFICATION',
    TemplateTypeWelcome: 'TEMPLATE_TYPE_WELCOME',
    TemplateTypeNewsletter: 'TEMPLATE_TYPE_NEWSLETTER',
    TemplateTypeAlert: 'TEMPLATE_TYPE_ALERT',
    TemplateTypeReminder: 'TEMPLATE_TYPE_REMINDER',
    TemplateTypeConfirmation: 'TEMPLATE_TYPE_CONFIRMATION',
    Unrecognized: 'UNRECOGNIZED'
} as const;

export type TemplateType = typeof TemplateType[keyof typeof TemplateType];